test: merge discount-status tests into existing HibernateBillDAOTest
